requirements.txt
Django==1.8 
mongoengine==0.9.0 
pymongo==2.8
django-rest-framework-mongoengine==2.0.2
django-rest-swagger==0.2.9 
djangorestframework==3.0.5 
PyJWT==1.1.0
django-cors-headers==1.1.0
